% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_optimization.R
\name{lin_reg}
\alias{lin_reg}
\title{Linear regression using gradient descent method}
\usage{
lin_reg(
  response,
  predictor,
  alpha = 0.05,
  learning_rate = 0.1,
  iterations = 1000
)
}
\arguments{
\item{response}{A \code{vector} of dimension 1 that represent the response
variable. It is also known as the "y" variable.}

\item{predictor}{A \code{dataframe} that contains all the predictors to
perform the regression. It is also known as the X variable.}

\item{alpha}{A \code{numeric} (continuous) used to denote the level of
significance when performing the confidence intervals. The default value is
0.05. The alpha parameter must be between 0 and 1.}

\item{learning_rate}{A \code{numeric} (continuous) used to control the
algorithm rate of learning. The default value is 0.1.}

\item{iterations}{A \code{numeric} (integer) that represents the number of
iterations that the gradient descent will be running. As this value
increases the error will reduce.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{Betas}{Regression coefficients}
     \item{Confidence interval}{Confidence interval for all the predictors}
     \item{Cp}{Cp-mallows}
     \item{F statistic}{Statistic F to compute the p-value}
     \item{P-value}{p-value}
}
}
\description{
Perform a linear regression using the gradient descent
method to find the coefficients of the regression. The user can
indicates the learning rate and the number of iterations.
}
\examples{
lin_reg(response = x, predictor = y, alpha = 0.05, learning_rate = 0.1,
iterations = 1000)
}
\author{
Group 7
}
